chore: split off util functions
